+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [1.0.2] - 2022-04-04
 
 - first commit
+
+## February 28, 2023
+
+- add create ticket command (Might need to re-login to Linear in VSCode)